Refactor error checking to service layer and add controller error handling

+++ src/main/java/ie/philb/springtodo/service/impl/TodoServiceImpl.java
@@ -6,9 +6,14 @@ package ie.philb.springtodo.service.impl;
 
 import ie.philb.springtodo.domain.Todo;
 import ie.philb.springtodo.domain.User;
+import ie.philb.springtodo.exception.TodoAccessException;
+import ie.philb.springtodo.exception.TodoException;
+import ie.philb.springtodo.exception.TodoNotFoundException;
+import ie.philb.springtodo.exception.TodoStateException;
 import ie.philb.springtodo.repository.TodoRepository;
 import ie.philb.springtodo.service.TodoService;
 import java.util.List;
+import javax.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,12 +31,34 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
-    public Todo getTodoById(long id) {
-        return todoRepository.getById(id);
+    public Todo getTodoById(long id, User user) throws TodoException {
+
+        Todo todo = null;
+        User owner = null;
+
+        try {
+            todo = todoRepository.getById(id);
+            owner = todo.getOwner();
+        } catch (EntityNotFoundException ex) {
+            throw new TodoNotFoundException(id);
+        }
+
+        if (owner.getId() != user.getId()) {
+            throw new TodoAccessException(id, user);
+        }
+
+        return todo;
     }
 
     @Override
-    public Todo save(Todo todo) {
+    public Todo save(Todo todo, User user) throws TodoException {
+
+        if (todo.getId() != 0) {
+            // Validate original, if it's not found or doesn't belong to us
+            // it will throw an exception
+            Todo original = getTodoById(todo.getId(), user);
+        }
+
         return todoRepository.save(todo);
     }
 
@@ -41,10 +68,14 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
-    public void delete(Todo todo) {
+    public void delete(Todo todo, User user) throws TodoException {
+
+        if (todo.getOwner().getId() != user.getId()) {
+            throw new TodoAccessException(todo.getId(), user);
+        }
 
         if (!todo.isComplete()) {
-            throw new IllegalStateException("Cannot delete todo with status " + todo.getStatus());
+            throw new TodoStateException(todo);
         }
 
         todoRepository.delete(todo);
